remove search2.go
